Rename clock slice files
- Preparation for refactoring clock slice
