+++ src/actions/live.ts
@@ -3,7 +3,7 @@ import { LiveGames } from '../models/live.js';
 import {
   LIVE_HYDRATE
 } from '../slices/live-types.js';
-import { ClockState } from '../slices/live/clock-slice.js';
+import { ClockState } from '../slices/live/clock-reducer-logic.js';
 import { ShiftState } from '../slices/live/shift-slice.js';
 
 export interface LiveActionHydrate extends Action<typeof LIVE_HYDRATE> { gameId?: string, games?: LiveGames, clock?: ClockState, shift?: ShiftState };
